backend/requirements.txt
fastapi
uvicorn
pandas
numpy
requests
python-dotenv
motor
apscheduler